Replace TEST_CASE with TEST_SUITE for tests. (#2075)
Newer versions of googletest prefer TEST_SUITE over TEST_CASE.

On the `Kokoro Windows` build we cannot use vcpkg to install googletest
because vcpkg has an old version of googletest. So we switch to using an
external project *just for googletest*. A bit ad-hoc, but will do while
we try to get googletest to create a release and so forth.

Manually remove the googletest vcpkg.

The package is sticky in the cache. Have to manually remove the
installed files or they are used instead of the external project
version.