Rename format_rfc3339.{h,cc}. (#2101)
As part of the changed for signed URL V4 we are going to need other
functions to format `std::chrono::system_clock::time_point`. This change
is a pure refactor, so the file name makes sense when I add more
functions to it. Oh, and I want a single file because there is some
platform-specific stuff in the `.cc` file that I do not want to sprinkle
around.