Cleanup handling of SignedUrlRequest options.
I think it is cleaner for the options to perform no translation on their
inputs, and let the request class do any processing. This will be more
useful with the V4 signed url request, where the handling of request
headers is slightly different.

clang-tidy was unhappy with `set_option()` for a long function name, so
I changed all of them to `SetOption()`, they are all private anyway, so
no harm.