A build for google-cloud-cpp as dependency.
Some of our customers will use google-cloud-cpp as a Bazel dependency,
there is Starlark code that loads `google-cloud-cpp` and its
dependencies, and this code should be tested.

We currently have a build for this on Travis. This creates the
configuration files for Kokoro, the plan is to remove the Travis build
once the Kokoro build is working.