Support different RSA-SHA256 signature formats.
For V4 signed URLs we need to encode the signature using hex format,
while for V2 signed URLs we need to encode them using Base64. In this PR
we decouple the function to create the signature from the functions to
encode them.

I changed the return value of SignStringWithPem to signal that it is a
byte array. In a separate PR I am planning to cleanup all these
functions because they are handling byte arrays and strings kind of
loosely. That is a cleanup independent of the V4 signed URLs though.