Implement V4SignedUrlRequest. (#2128)
This class in not part of the public API, it will be used by
`storage::Client` to accumulate all the optional parameters in a V4
signed URL request. Just like we did for V2 signed URLs, this class just
creates the blob to be signed, the actual signing happens in the
credentials class.

I will also need to insert the `host` header if the application does not
provide one.