Remove OpenSslUtils struct. (#2141)
This "struct" only had static member functions. I just moved the private
member functions to the .cc file and things are copacetic.

With Bazel we use BoringSSL, which has a different API for Base64
encoding and decoding. Also moved the includes for OpenSSL headers to
the .cc file, because that is where we use them.

Refactor UrlsafeBase64Encode implementations.

Thanks to @devbww for the suggestion. Also refactored RightTrim()
because there was a single use of this function.