Refactor HmacKey requests. (#2199)
First, I noticed that we could get rid of the `Client::*ForProject()` overloads if we
used an option to override the default project id.  That reduces the overall surface
and fits more naturally with all the other optional parameters.

Then I noticed that there was a lot of commonality across all the request types for
HmacKeys. We only have two such types so far, but we will get 5 of them, all with
a common `project_id()` field and (more interesting) with special handling for
`OverrideDefaultProject`.