Fix build breakage.
I did not notice one of the PRs was going to break the build. Note that
it did *not* break the build, the build passed once, but the second run
would be broken because the object versioning examples changed the state
of a global bucket.

The "Right Thing"[tm] would be to not use a global bucket. I plan to do
that in a future PR, but for now we need to get the builds restored
quickly.