Fix Kokoro + Windows build. (#2222)
For some reason `vcpkg integrate` is freezing. We have not needed this
for a long time (when we switched to Ninja), so I just removed it.

I had to add some more debug logging to get here, which I think it is
generally useful, so I am leaving it in.