Initial draft of "flags" for integration tests.
The integration tests have been using this weird pattern where we store
the command-line flags as static member variables of the test
"Environment", and then we initialize these member variables in the
environment constructor.

This is a first draft of how we could stop doing that. Ideally we would
use a real command-line parsing library (absl::Flag is not a thing,
Boost.ProgramOptions is a depedency hassle).