Disable signals in libcurl. (#2248)
TIL: libcurl uses signals to perform timeouts by default. This change
disables the signals for all the handles we create, per the
documentation on:

https://curl.haxx.se/libcurl/c/CURLOPT_NOSIGNAL.html

I will try to write a regression test later, though it might be hard
because it requires "waiting for a couple of minutes" to get the
signal.

In any case, given the documentation, and the fact that there is a bug
report (#2225) associated with this, I believe we should incorporate
these fixes.