Implement asynchronous retry loop for simple RPCs. (#2252)
This PR implements an asynchronous retry loop for simple (unary) RPCs.

The loop calls `CompletionQueue::MakeAsyncRpc`, on retryable failures it
sets a timer (based on the `RPCBackoffPolicy`) and tries again when the
timer expires.  If the `RPCRetryPolicy` is exhausted, or if the operation is
non-idempotent, or if the error is not retryable then the operation fails, and
the last error is reported back to the application (via an `future<StatusOr<T>>`).

In addition to implementing a unit test for the retry function, I modified
`TableAdmin::GetTable` to serve as a template.