Write a custom module for find protobuf.
Finding protobuf is a mess (see the documentation in the module for
details), but with this module we can just type:

find_package(ProtobufTargets REQUIRED)

and it works. In the future we will use this module to remove the
googleapis submodule, and we will install it so application developers
can benefit from a simpler way to discover protos.