Use standard name for grpc_cpp_plugin target. (#2270)
gRPC creates a gRPC::grpc_cpp_plugin target, so why create a target with
a different name for the same thing? Note that sometimes we need to
create the target ourselves, maybe because the gRPC library does not
have CMake-config files installed. But even then, we should just use the
standard name.