Add a FindgRPC module. (#2276)
With this module we can simply write:

```
find_package(gRPC REQUIRED)
```

and this works whether gRPC was installed via CMake or via Make. This is
needed to remove the googleapis submodule, because when compiling
googleapis we want a simple way to discover where gRPC is. It should
also help us simplify the INSTALL instructions for customers that want
to use CMake.