Fix openSUSE build.
openSUSE's version of libcurl returns different error codes when
trying to connect to an invalid URL (we use http://localhost:0 in our
tests). Changed the curl_client_test to just check for "error" and not
care about the actual error code.

Also added a separate test for the CurlHandle::AsStatus() function.