Enable presubmit for "test-install" builds. (#2294)
I thought we could away without these, but evidently not. A weird
platform specific bug just sneaked by (the openSUSE:Leap build is
broken), and I am about to submit changes to test the installed CMake
modules, and that is very tedious to test without these presubmits.

Note that merging these files will not active the builds, this is
just in preparation for a future change in the internal configuration
for Kokoro.