Workaround build problems on openSUSE:Leap.
This weekend openSUSE:Leap released the 1.1.0i version of the OpenSSL
runtime libraries, but they did *not* release the development libraries
at the same time. Effectively this means that installing the development
libraries requires a downgrade.

I just changed both openSUSE:Leap and openSUSE:Tumbleweed to always
allow a downgrade when installing the development tools.