Make multiple attempts to download json.hpp.
Some of the CI builds are failing because this file fails to download.
Use the standard technique: make multiple attempts to download, using an
exponential backoff.

Typically we rely on ExternalProject_Add() to implement the download
retry loop, but for this file we cannot because early versions of
ExternalProject_Add() cannot download single files.