Change CMake-config files to work without `pkg-config` dependency. (#2323)
In this PR we fix `FindgRPC.cmake` to do not depend on `pkg-config`, and change the
generated config files for `googleapis` and `bigtable_client` to take advantage of this
new file.

To test these changes I modified the `ci/test-readme/Dockerfile.*` files to also build
sample programs using CMake after installing `google-cloud-cpp`. Separately we enabled
these builds as presubmit tests.