Update reference ABI after latest changes. (#2336)
I recently changed some data structures in google/cloud/storage. I
thought it would be interesting to highlight how these changes affected
the ABI and (to the degree it is measurable) the API:

https://source.cloud.google.com/results/invocations/b5597a24-982c-45ec-b9e9-945a7c9ddc8d/targets/cloud-cpp%2Fgithub%2Fdocker%2Fcheck-abi/log

Starting on line 4363 (sorry, can't create direct link). In summary the
changes that affect "Source Compatibility" are:

- I removed some functions from the `google::cloud::storage::internal::`
  namespace. This change is allowed, though technically API breaking.
- The name of a field in `internal::CreateHmacKeyResponse` changed, this
  also allowed, though technically API breaking.

The report also highlights a number of A*B*I breaking changes, which I
list here for completeness sake, though we allow these at any time:

- The size of the `HmacKeyMetadata` class changed. Note that the
  report does not call this an API breaking-change.
- There is a similar change to `internal::HmacKeyResponse`.