Use BulkApply to cleanup the test table.
Before running each integration test we cleanup the test table, removing
all rows from it. We used `DropAllRows()` for this, but there is limited
quota for this operation and the builds are failing because we exceed
that quota. Use `BulkApply()` when the number of rows is relatively
small, which is most of the time, and fallback on DropAllRows()
otherwise.