Workaround a bug in the bigtable emulator.
The version of the emulator we can use has a bug in dealing with rows
deleted via Apply() or BulkApply():

https://github.com/googleapis/google-cloud-go/issues/1240

The bug is fixed, but that version has not made it to a usable SDK yet.
The bug does not affect rows deleted with DropAllRows(). So to save
DropAllRows() quota we will use BulkApply() when running in production,
and DropAllRows() when using the emulator.

Since we need to know whether we are using the emulator in the
TableIntegrationTest fixture, refactor the function to check that.

Also fixed a typo in the generation of the random instance ID for tests
against the emulator.