Fix flakiness caused by drop all rows quota (#2355)
Before running each integration test we cleanup the test table, removing
all rows from it. We used `DropAllRows()` for this, but there is limited
quota for this operation and the builds are failing because we exceed
that quota. Use `BulkApply()` when the number of rows is relatively
small, which is most of the time, and fallback on DropAllRows()
otherwise.

This basic plan became a bit complicated because the version of the
emulator we can use has a bug in dealing with rows
deleted via `Apply()` or `BulkApply()`:

googleapis/google-cloud-go#1240

The bug is fixed, but that version has not made it to a usable SDK yet.
The bug does not affect rows deleted with `DropAllRows()`.

So to save `DropAllRows()` quota we will use `BulkApply()` when running
in production, and to workaround the bug use `DropAllRows()` when using
the emulator.

Since we need to know whether we are using the emulator in the
`TableIntegrationTest` fixture, refactor the function to check that.