Do not raise exceptions in ReadObject().
The plan was to return the status in the ReadObjectStream, that was not
working when there were errors (such as authentication problems)
preventing us from even creating a valid ReadObjectStreambuf. The
"solution" proposed here is to create a streambuf that is permanently
set to "this is an error stream, everything is terrible".