Report errors detected in CurlDownloadRequest. (#2378)
CurlDownloadRequest was dropping errors on the floor, when it should
not. I think I did not do a great job converting this class to StatusOr.
In any case, I added an integration test to show that the errors are
reported. Note that the errors are not handled correctly yet (no
retries for example), that will be the subject of a future PR.