Make `StartAsyncPollOp` accept future<Operation>.
Per coryan@'s suggestion, StartAsyncPollOp is always used in then, se we
might make it friendlier to use in such a case.

This change allows calling StartAsyncPollOp before the operation is
ready, which allows us to not capture arguements required for its
creationg in yet another entity - that was cumbersome.