Multiple fixes to the HMAC implementation.
The `serviceAccount` query parameter is now called
`serviceAccountEmail`.

The `secretKey` field when creating a HMAC key is now called simply
`secret`.

The testbench was not returning errors correctly, and some errors did
not include the necessary debugging information.

The UpdateHmacKey() request should use `PUT` not `POST`.

Only inactive HMAC keys can be deleted, so the testbench needs to check
this, and the unit + integration tests should deactivate a HMAC key
before trying to delete them.

DeleteHmacKey no longer returns the metadata for the deleted HMAC key,
it returns an empty response (which we represent by `Status`).

Set Content-Length in CreateHmacKey(). For somewhat silly reasons
`CurlRequestBuilder` does not automatically set the `Content-Length:`
header to 0 when the payload is empty: the same code path handles POST,
PUT, GET, and DELETE requests, and the header is only required for POST.