Fix flaky HMAC key integration tests.
I was being paranoid about leaving HMAC keys undeleted after a test, so
I introduced loops to delete them at the end of the integration tests.
But that is racy because another build may be trying to run the tests at
the same time and their keys would be deleted too.

We should consider creating a new service account for each build, so
they are better isolated from each other, but that requires additional
permissions, I created #2435 to track that.