Refactor UsingTestbench() function.
In some tests we need to know if we are running against the testbench,
for example, because the testbench implements special headers to control
its behavior, or because the testbench does not implement some
functionality (signed URLs is the one I can think of). We had multiple
versions of this function.