Cleanup implementation oauth2::*Credentials.
I needed to implement ComputeEngineCredentials::AccountEmail, a virtual
function to determine the service account associated with a Credentials
object, if any.

In the process I started a small exercise in yak shaving. There is some
(rather trivial actually) code refactored from 3 oauth2::*Credentials
classes into oauth2::RefreshingCredentialsWrapper. There was a lot of
coupling in this code, and the const-correctness had (and probably still
does have) much to be desired.

I actually stopped myself before doing far more cleanup, because what I
need for the SignBlob support is fixed, and the code is better, if not
quite as good as it could be *yet*.