Add SigningAccount options to Sign*Request. (#2447)
When singing a blob, either as part of a signed policy document or a
signed URL, we may want to use a different account (and it must be a
service account) than the account used for authentication. This is the
only way to use a authorized user account to access GCP, but sign the
URLs with a different service account. Or when using GCE authentication
we may want to use a different service account for the signed URLs.

Note that these options are currently unused, a future PR will connect
them to the rest of the code.