Use SFINAE to constrain applicability of the BulkMutation(M&&...) ctor
The unrestricted BulkMutation perfect-forwarding constructor was a
better match than the (generated) copy constructor for a non-const
lvalue, so its embedded static_assert requiring convertibility to a
SingleRowMutation meant that non-const BulkMutation lvalues could
not be copied (so they could not be used as pass-by-value function
arguments, for example).

So, make the "convertibility to a SingleRowMutation" requirement a
SFINAE constraint instead, meaning the perfect-forwarding constructor
will be disabled when it does not apply, and that allows the copy
constructor to be found instead.