Moves enablers to LHS of assignment to prevent callers specifying.
It's preferable to use SFINAE on the left-hand side of the assignment so
that it's declaring the type of a non-type template parameter rather
than on the right-hand side where it's instead setting the default type
for a regular (type) template parameter. This way is preferred because
it prevents callers from overriding the SFINAE by explicitly specifying
the template parameter, which callers could do if the enabler were only
setting the default.