Wrap streaming read RPC loop. (#2533)
This introduces a new way to implement the even loop for a streaming
read RPC. This version uses callables (as opposed to pointers to member
functions) to wrap the call that creates a asynchronous streaming read
RPC, I think those are more readable (and allegedly more efficient).

This version also uses the `PrepareAsync*()` generated functions, which
are less racy than `Async*()`. The latter immediately start the loop by
calling `Start()` on the streaming read RPC, which can create a race
because the most common action in response to a successful `Start()` is
to call `Read()` on the returned `ClientAsyncReader<>`, but said reader
may not have been returned yet.