Fix noex build and improve build logs. (#2539)
We cannot install `clang-format-7` on all versions of Ubuntu. We already install it on the
section of the Dockerfile for Ubuntu:18.04, so remove it from the generic section, and things
are happy.

I also changed the CI build scripts to actually print any errors while creating the docker images.
I could not figure out why the build was broken until I did that.