Return google::cloud::future from InstanceAdmin functions.
A few functions were returning `std::future<>` because they were implemented
before we had `google::cloud::future<>`, changed the return type to
be consistent with the `Async*()` variants.

Changed the implementation to use `Async*()`, that seems simpler than having
an implementation using blocking calls in a separate thread.

Removed a bunch of redudant unit tests once the implementation
is refactored. We already have unit and integration tests for
the `Async*()` versions of these functions.