Prevent transient problems from breaking production test.
One of the test runs against production failed in this test:

https://source.cloud.google.com/results/invocations/7e17585e-596f-40c3-b237-e4745b54dfc1

The log is not detailed enough to diagnose the problem, so first I
fixed that. My guess is that the test failed because the mutations
used to prepare data for the test are non-idempotent, so any transient
failure will stop the test. I fixed that too.