Remove polling loops from samples with future.
Some of the code snippets showing how to use functions that return
`google::cloud::future<>` did a loop polling the status of the future.
This takes a lot of vertical space in the code sample, and does not add
much. This PR makes these examples simpler, and still show how to poll.