Fix the admin tests and examples to run against production.
Introduce new scripts to run the admin integration tests and examples
against productions. There were a number of small problems in the
tests and examples that I had to fix to get the tests to pass.

These tests will be enabled in a future PR as part of a nightly build.