Update documentation for top-level classes.
I updated the doxygen documentation for bigtable::Table,
bigtable::TableAdmin, and bigtable::InstanceAdmin. Fixed a lot of
nits, documented which operations are idempotent (and thus normally
retried), documented how to configure the policies in bigtable::Table,
their default values for all 3 classes. And also which operations are
thread safe on these classes.