Avoid SIGPIPE signals when connections break.
Automatically setup a signal handler for SIGPIPE on platforms where it
is needed. libcurl recommends setting CURLOPT_NOSIGNAL for
multi-threaded applications (and we expect the library to be used in
such applications):

https://curl.haxx.se/libcurl/c/threadsafe.html

"When using multiple threads you should set the CURLOPT_NOSIGNAL option
to 1L for all handles."

"Note that setting CURLOPT_NOSIGNAL to 0L will not work in a threaded
situation as there will be race where libcurl risks restoring the
former signal handler while another thread should still ignore it."

But also note that when setting CURLOPT_NOSIGNAL to 1L the documentation
recommends:

"When CURLOPT_NOSIGNAL is set to 1L, your application needs to deal with
the risk of a SIGPIPE (that at least the OpenSSL backend can trigger)."

This change implements this last recommendation.