Implement MutationBatcher without using noex::Table. (#2565)
This version intentionally drops some functionality, specifically the
previous version was able to handle partial results from a BulkApply()
to admit more mutations early. Unfortunately the server does not send
partial results yet, and the code is more difficult to maintain.

As @dopiera suggested, we can simplify the data structures now that
MutationBatcher is non-streaming.