Fix support for empty chunks in resumable upload. (#2666)
* Fix support for empty chunks in resumable upload.

When doing resumable uploads we may need to send an empty chunk at the
end of the stream. For example, because the stream has a round number of
chunks, and all of them are sent before the stream is closed. At that
point we need to send some indication to the resumable upload session
that the upload is "done", and an empty chunk is the only choice.

Our support for this was broken, which was Okay-ish because we do not
use it (yet): our streams almost always have data in the buffer (that is
a separate bug), and flush this buffer in the last request.

With this change we fix the formatting of the Content-Range header to
support empty chunks, add an integration test to verify this works
against production, and fixed the testbench to parse these chunks
correctly.