Refactor common code in storage/benchmarks. (#2661)
I am about to create additional benchmarks for the GCS library, so it is
time to refactor common code across these benchmarks. That requires a
new library, some unit tests for the library functions, and a lot of
additional complexity in the BUILD and CMakeLists.txt files to support
both.