Change `AsyncRowReader` deletion semantics.
Before this PR, AsyncRowReader existed until all the rows were read from
it or until a permanent error occurred.

This PR enables the user to drop references to `AsyncRowReader`
mid-stream, which will result in the reader cancelling the stream.

This is essential, because it allows implementing `AsyncReadRow` via
`AsyncReadRows` in the most trivial way:
```C++
future<StatusOr<optional<Row>>> AsyncReadRow(...) {
  return AsyncReadRows().Next();
}
```