Send larger chunks is CurlResumableStreambuf. (#2668)
The class was sending chunks of size 256KiB (the default configuration parameter),
even when the application provides a multi-megabyte buffer, and then it would save
the buffer to send it later, on the small 256KiB chunks. Sigh...

This version still performs a lot of data copying, but that is a matter for a future
change (there is already an open bug for it).

The iostream classes keep pointers to a "buffer", which we must own. We need
to keep the current_ios_buffer_ buffer in sync with these pointers, and we
were not doing it right for data inserted one character at a time.