cleanup!: Use google::cloud::future.
BREAKING CHANGE: Fix the last function that returned `std::future` to
return `google::cloud::future<>`. The function name changes too, to
be more consistent with similar functions.

I removed the unit tests because (a) the tests would only be checking that
WaitForConsistency() calls AsyncWaitForConsistency(), and (b) because we
already have converage via the integration tests and example.