bug: Remove CURL* handle from multi-handle on errors. (#2688)
Otherwise the CURL library is left in an inconsistent state. This is part
of the work for #2655.

With the more complex code path to remove the handle I need to explicitly
track if it is in or not.  I think there is something we are doing
sub-optimally here, but right now I have a bug to fix.

This change adds a new member variable in
storage::internal::CurlDownloadRequest. That is a ABI-breaking change,
so we need to update the baseline. Note that the checker correctly
reports this as non-breaking for the API.