feature: Implement option to read from an offset. (#2690)
Allow application developers to read an object starting at a given
offset. This will also be used to restart downloads from the last
received byte. Fix the code to properly combine this option and the
gcs::ReadRange() option: the highest of the range beginning and the
offset is used as the starting point.

Refactored some code to avoid duplication.