bug: Service account credentials not refreshing properly. (#2691)
The code was not updating the expiration time when creating a new credential,
effectively "refreshing" by creating an immediately expired credential. Most of the
changes are (as usual) needed to make the code testable, in particular,
`oauth2::ServiceAccountCredentials` was not using the mockable clock everywhere,
and we needed functions to parse url-safe-base64-encoded strings.

I updated the API baseline because there are a number of changes to private
implementation classes, but the API/ABI checks do not know what is public vs.
what is private.