feature: Automatically restart downloads on error. (#2693)
A download can be interrupted for a number of reasons, but most commonly
because the application does not read data from a download for several
minutes (about 6 seems to trigger a disconnect, but that is purely an
observation). With this change the download is restarted from the last
received byte, and using the object generation used in the original
download too.

This change updates the API/ABI baseline because many classes in
`google::cloud::storage::internal` changed. Since this namespace is not
part of the public API these do not need to be flagged in the release
notes, but the tools we use flag all changes.