Reduce data copies in ObjectReadStream.
When the application uses `std::basic_stream<char>::read()` we can copy
the data received in the download straight into their buffer. And when
the application calls something like `getc()` we can still copy directly
into the internal buffer in the `ObjectReadStreambuf`.

I had to change the internals of `CurlDownloadRequest` to support this:
libcurl requires write callbacks (this is the callback for downloads,
strange naming, but whatevs) process the full buffer they are given.
When the application gives us a buffer in `.read()` that does not line
up perfectly with the write callback buffer we need to save the data
somewhere. I added a `spill` buffer in `CurlDownloadRequest` to keep
this data. These are typically small and we can use a fixed-size array
to support them.