Fix build for macOS+CMake.
I did not think about macOS when using RUSAGE_THREAD, which is not a
thing on BSD-based systems. Fixed the code to detect if the symbol exists,
if it does use it, if it does not, do not allow more than one thread in
the benchmark that measure CPU utilization.