Update googleapis to a recent version.
Newer versions of googleapis include Bazel BUILD files to compile (most)
of the protos, and in many cases include Bazel rules to generate C++
libraries for these protos. Unfortunately we still need a custom
googleapis.BUILD file because the bigtable protos do not have C++
support (yet).

I had to update gRPC at the same time, because it includes better
`cc_grpc_library()` rules too.

One has to be careful to disable the support for other languages in the
WORKSPACE file, or the googleapis Bazel rules require code generators
for Java, Go, and possibly other languages.

I was also able to simplify the dependency on OpenCensus, as gRPC now
pulls the right dependencies for it.

Note that I removed the support for the Spanner protos in the
googleapis.BUILD

This is all based on work that @scotthart did for the Spanner library.