bug: Use one mutex per data kind in CURLSH. (#2730)
We are using libcurl incorrectly. I misread the documentation for
libcurl and created a single mutex for all the "kinds" of data in a
CURLSH handle. This is incorrect:

https://curl.haxx.se/mail/lib-2019-06/0011.html

I changed the library to use one mutex per data kind. I wrote a
program to reproduce the bug, it never worked, but I think the
documentation on how to run large-scale tests is useful, so I
am merging the program anyway.