Adds support for spanner ARRAYs. (googleapis/google-cloud-cpp-spanner#58)
A Spanner ARRAY is represented as a std::vector, where T is any other
valid Spanenr type. Well, technically, arrays of arrays are not allowed,
but we don't currently forbid this. We could/should add an enabler to
this constructor to prevent this.

This PR also changes op== to do comparisons by looking at the data member protobuf contents instead of converting to C++ types and using their op==. This change makes it easier to recursively call the equality function for containers that can contain other values and containers.

Closes googleapis/google-cloud-cpp-spanner#53