Move libc++ build to Bazel. (#2735)
Use Bazel to compile with libc++, this is much easier to maintain than
the CMake approach as Bazel compiles all the dependencies from source
with the same flags.

One caveat: Bazel injects `-lstdc++` always, even though the compiler
does not need it. I was unable to disable this, but seems harmless as
`-lc++` and `-lc++abi` appear first in the command-line. Also, I
verified only the libc++ headers are included (`__GLIBCXX__` was not
defined).