Make build metadata configurable.
On development builds we annotate the build with the short git commit
id, but in some cases this is not available (Google Cloud Build does not
send the .git directory to the server), and in other cases we want to
override it. Allow the developer (or script) calling `cmake` to set the
build metadata.

Update API/ABI baseline as this change removes a function in
`google::cloud::internal::`.