Change hash validators to operate on raw buffers.
To minimize data copies we need to operate on `char*` and `char const*`
buffers provided by the application without storing them in a
`std::string`. This PR changes the hash validators to work with such
buffers.

Update the API/ABI baseline because we changed the API for classes in
`google::cloud::storage::internal::`.