Use check_cxx_symbol_exists() to search for RUSAGE_THREAD. (#2752)
We were using check_symbol_exists() which uses the C compiler. On Linux, the C++ compiler defines `_GNU_SOURCE` by default, that enables `RUSAGE_THREAD`, and we can use it in our code. The CMake test based on the C compiler could not find the symbol because the C compiler does not define `_GNU_SOURCE`.