Disable optional's converting c'tor when U == optional. (#2757)
This change disables optionals converting constructors (i.e.,
`optional(U&&)`) when `U` is the same as `optional<T>`. In this
situation we instead want to use optional's normal copy constructor.

This bug showed up in our gcc 4.8 tests when using `optional<bool>`. I
*think* gcc 4.8 exposed this issue because it probably *copies* the
optional bools (e.g., on resize) instead of moving them. And an lvalue
of type `optional<bool>` *is* explicitly convertible to type `bool`
because of its `explicit operator bool()`, so the converting constructor
was in the overload set, and in fact it was a better match for a
non-const lvalue than was the copy constructor whose argument type was a
const value.