Reject IAM Binding protos with unknown fields. (#2764)
The key motivation is the `condition` field, which breaks the assumption
that the there is one member list per-role. However, to make it a tiny
bit more robust we reject all Binding`s unknown fields.

As a byproduct, parsing JSON was changed to return Status instead of
throwing (it had been doing both).