bug: Treat files as binary for {up,down}loads.
The storage Client::{Upload,Download}File functions should treat the
underlying files as binary. We should not be in the business of parsing
the customer data. If the customers want to convert the files from
Windows format to Unix format during the transfer they can use
`Client::WriteObject()` and `Client::ReadObject()` instead.