chore: Minimize impact of package server outages.
When the package servers for a distro go bad (and they do from time to time),
our builds break because we cannot rebuild the images with the latest and
greatest packages. In this PR we use Google Container Registry (GCR) to cache
the Docker images with all the base packages we need to build the code.
We do this only for the *install* tests, because those use the more obscure
distributions, and they tend to fail more. We will use the same techniques
for all builds in a future PR.

The expectation is that GCR has a better observed SLO that the combination
of all the package servers out there. It is based on GCS, so there is good
reason to believe it should work well.

To prevent our base images from getting stale, the builds on the `master`
branch:
- Try to build the base image, without using the cache.
- If successful, they push the new base image up, and use that to run
  the build.
- If not successful, they still use the cached image to run the build.

This technique does not work well if the Dockerfile installs additional packages
after the initial base image. I plan to refactor the builds in a separate
PR to eliminate such cases.

The cache is also invalidated when the contents of the base image changes,
that is, we add a new tool or depedency to the image. That is relatively
rare. We expect that rolling out such changes when the package servers
are having an outage does not happen often, and if it does we will just
need to wait the outage out.