Automatically build image for benchmarks.
In this PR we describe how to automate the steps to create the Docker
image for the benchmarks. Once this is merged to `master`, we can enable
the benchmarks and save their logs, but that will come in separate PRs.

Also include the bigtable benchmarks in the image, as we will enable
those in a future PR too.