Initial config for Bigtable benchmarks.
Create GKE configuration file (templates) to launch the Bigtable bencharks.
The benchmarks run every hour, using the latest version of the Docker image available
at that time. This Docker image is automatically created on each commit to `master`.

We run the benchmarks periodically because this is the easiest way to guarantee that
only one instance of each benchmark is running.

The raw logs from the benchmarks are captured in BigQuery tables, the analysis and
dashboards will come in a future PR.