Run the storage benchmark as a CronJob. (#2790)
Using a cron job seems to work better: at most one copy of the benchmark
is running at a time, and it picks up the new Docker image without
having to reset the YAML file.

Also save the raw logs to BigQuery. It is "easy" to then parse the logs
or output them to a more structured BigQuery table (and then plot stuff
there).