Reduce import friction into Google. (#2794)
When we import the code into Google we need to apply a number of code
transformations. The least desirable of these transformation is a series
of patches to handle how some `bytes` proto fields are mapped to C++
inside Google. These fields are always mapped to `std::string` in the open
source version of protobuf, but are handled as a different type, inside Google.
For extra fun, this type does **not** have the same API as `std::string`.

This PR introduces three changes:

- A type alias for each of the affected fields: `RowKeyType`, `ColumnQualifierType`
  and `CellValueType`. It is important to note that we have **no plans** to change
  the definition of these aliases for the external version of the C++ client.  These will
  map to `std::string` for the foreseeable future. Changing the definition of these would
  be considered a major API breakage.
- Overloads for most constructors (and several functions) that can accept `std::string`
  or the internal Google type. As a side benefit, these overloads also accept
  `std::string_view` when you are using C++17.
- A series of functions (in `google/cloud/bigtable/internal/google_bytes_traits.h`) to
  operate on these types. Inside Google we will have different versions of these
  functions.