feat: Install .proto files for googleapis-c++ libs.
When we compile the googleapi libraries we should install their proto
files too, that way other protos that import them may use them too. I
refactored the code to generate these libraries because it was really
tedious to change them otherwise. Note that the destination directory
for these proto files is $PREFIX/include, the protobuf library places
its own protos there, so I figured it was the right place.