cleanup: Simplify generation of version_info.h.
These files used to be generated by CMake and BUILD files on each
compilation, but they rarely change, and to support the BUILD generation
we generated and committed *_version.bzl files with basically the same
contents. In this PR we change the CMake scripts to generate
version_info.h in the source directory. We commit these files and the
BUILD files are simpler. This also removes the need for those
*_version.bzl files.