Added ToProto/FromProto support for Value (googleapis/google-cloud-cpp-spanner#131)
Adds support for getting the raw protos from a spanner::Value and constructing a spanner::Value from raw protos. This is needed for converting the raw request/response protos to/from spanner::Value objects.

Note: In a later PR, we'll add support for checking for invalid protos. See googleapis/google-cloud-cpp-spanner#122